Add fetch tool for retrieving Chromium's custom-built Rust toolchain
Adapt the approach [1] that Chromium uses for fetching its Rust
toolchain to Skia. Take update_rust.py and the minimally required helper
functions from Chromium's clang update script to fetch a specific, fully
functional and self-contained Rust build from Chromium's CDS bucket. In
these buckets, Chromium stores its Clang and Rust toolchain builds.

Maintaining and building a tip-of-tree up-to-date rust toolchain
requires a level of constant effort that can be avoided on the Skia side
if we rely on Chromium's work here.

Currently, this script needs manual rolling to the respective latest
build in [2]. Issue skia:14191 tracks changing that so that the script
is upgraded to be able to automatically roll to the latest known-good
revision.

[1] https://source.chromium.org/chromium/chromium/src/+/main:tools/rust/update_rust.py
[2] https://commondatastorage.googleapis.com/chromium-browser-clang/index.html?path=Linux_x64/rust-toolchain-

Bug: skia:14185
Change-Id: I66657a4943251b15d1c2ce69044bbfbb5403b4d4
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/657157
Reviewed-by: Kevin Lubick <kjlubick@google.com>
Commit-Queue: Dominik Röttsches <drott@google.com>