[skottie] read in slots made with EssentialProperties on export
This is a replacement function for properties with a slotid, so all default property info is dumped if a slot is present.
There was talk of doing an intelligent "merge" for slots, specifically for text properties. We can pivot to the merge solution in a follow up.

Change-Id: I835910cc8408cce68d1a98d271f9d7d2537c8b90
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/659436
Commit-Queue: Jorge Betancourt <jmbetancourt@google.com>
Reviewed-by: Florin Malita <fmalita@google.com>