Enforce IWYU on GrDirectContext
A small win here is removing GrBackendSurface from a header,
which currently brings in all the chosen backends, making
the modularization effort harder.

In order to do so, I had to move some of the "convenience"
methods from the header to the .cpp.

I removed the INHERITED typedef because it is 1) no longer
in the style guide and 2) made IWYU recommend inclusion of
some priv file from src/

I chose to tell IWYU to keep the forward declared backend structs
instead of using #ifdef because it felt a bit strange to
do so (forward-declares are ~free).

Change-Id: Ided96a2a517c779514a4eb69d4bf0e79fe3a85c7
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/656436
Reviewed-by: Greg Daniel <egdaniel@google.com>
Commit-Queue: Kevin Lubick <kjlubick@google.com>