Reduce the size of the inverse table
We can use the fact that the table is symmetrical around zero
to reduce the number of entries.

I kept the negative side because it has one more entry
because 2s compliment number range is asymmetrical even
though there is an assert that forces symmetry.

Change-Id: I7a3bcb484ed917b6a8f19b779d8721ccb36337df
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/660857
Commit-Queue: Herb Derby <herb@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>