[skif] Add basic FilterResult test cases
This also formalizes the order of operations for the effect modifiers
on a FilterResult and fixes some discovered bugs in the applyCrop() and
applyTransform() implementations after the tests were written.

This CL defines some logic to run the filter result tests for the
Graphite backend but the synchronous readback isn't working correctly.
The graphite test definitions are commented out and will be re-enabled
when there's time to investigate. But the tests are valid for raster
and Ganesh so I don't want to block their landing.

Bug: b/264551694
Change-Id: Id89495dc44cbc84f6c83c3572f6b0546a7940978
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/612679
Reviewed-by: Robert Phillips <robertphillips@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>