[ganesh] Try harder to use SW renderer for paths
With this CL:
 - We try to use the SW path renderer for every path draw
 - We check if the SW path renderer can actually be used before trying
   to use it for a draw
 - We never try to use the SW path renderer with MSAA

This seems to have been the intent of the renderer selection code as of
http://review.skia.org/5774 in GrDrawingManager.cpp (based on comments
from http://review.skia.org/5763). If no path renderer was found to draw
with MSAA, we would try the entire chain again to draw without MSAA.
Later we removed this retry logic in http://review.skia.org/85043.

I don't know what the context was for change 85043, but my changes here
assume that it would have been better to keep the SW renderer in
consideration as a non-MSAA fallback when no path renderer is found to
draw with MSAA.

This has become more relevant in the context of
http://review.skia.org/631156, where we explicitly want to fall back on
the SW renderer for draws where we would normally use MSAA.

Change-Id: Ib8f17a529f821e09b9b56138d14b6a6e381cc4ab
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/658797
Reviewed-by: Greg Daniel <egdaniel@google.com>
Commit-Queue: James Godfrey-Kittle <jamesgk@google.com>