Remove FCI from public.bzl
No dependents need this anymore and it should no longer be the default.

Since there is no real sane default for this platform (users should
create the appropriate SkFontMgr at runtime and pass it around) attempts
were made to make the default factory `empty_factory` or
`custom_empty_factory`. `empty_factory` does not work as there are many
users that (somewhat reasonably) expect the default SkFontMgr to be able
to create fonts from data. `custom_empty_factory` does not work as there
are users expecting the default SkFontMgr to pick up on an FCConfig they
created and set as default in FontConfig with FcConfigSetCurrent. As a
result the default at the moment must be backed by FontConfig.

Change-Id: Id67d3f5db55fcc2e2154ac5a1c166d064ac68486
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/662219
Reviewed-by: Florin Malita <fmalita@google.com>
Commit-Queue: Ben Wagner <bungeman@google.com>