[graphite] Support all stroked lines in AnalyticRRectRenderStep
Stroked lines produce very similar geometry to stroked rectangles
except that they are not necessarily axis-aligned relative to the
SkShader's local coordinates. This is easy enough to address by folding
the line's direction into the Jacobian if necessary without trying
adjust a local matrix or construct an equivalent filled shape on the
CPU.

To support butt caps, this adjusts the vertex shader to support
outsetting on only one axis. This is used for both stroked lines and
for stroked rectangles that have a 0 width or height.

Adds a version of the existing strokedlines GM to use drawPoints
(e.g. drawLine) instead of drawPath that happens to be a line.
Adds an additional GM that was inspired by the strokerect GM that
shows the control point structure from an equivalent stroked path
and tests various tricky dimensions and tests local coord tracking.

Bug: b/274442707
Change-Id: Ifb2f72f2fcc4a37fe26806b9a6cd6ead9a5d514a
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/662198
Reviewed-by: Arman Uguray <armansito@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>