[graphite] BufferManager: Rename `mappable` and fix AccessPattern setting
Previously `AccessPattern::kGpuOnly` was only used when the platform
required the use of transfer buffers. On platforms where transfer
buffers are not required (e.g. Metal), this flag should take into
account whether the client intends to map the buffer (i.e. the
`mappable` flag is set to false). This enables GPU-only storage
buffers to be initialized with `MTLResourceStorageModePrivate`.

This also renames `mappable` to `supportCpuUpload` to better capture the
user's intent. While mappability/mapping is implied, this leaks an
implementation detail that is not entirely accurate when transfer
buffers are involved (where a mapped transfer buffer and an unmapped
GPU-only buffer pair gets created). The new naming only captures the
intent and not implementation of the buffer itself.

Similarly, getMappedStorage has been renamed to getStoragePointer. The
"mapped" naming is not incorrect in this context but getStoragePointer
mirrors its "getStorageWriter" counterpart.

Bug: b/243721614

Change-Id: Ied2ca18310d1f93ba0b0c6421bf9af0426fabfdd
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/665583
Reviewed-by: Greg Daniel <egdaniel@google.com>
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Arman Uguray <armansito@google.com>