Detect modification of the input color in ReturnsInputAlpha.
If the input variable has been written-to, then returning `input.a`
isn't sufficient to guarantee that alpha is preserved.

This fixes some Gold image regressions with the Overdraw filter.
It also required a bit of touchup in the HighContrastFilter, as this
code modified the input color's RGB while leaving its alpha alone;
our check is not savvy enough to determine this on its own, and needed
a helping hand.

Change-Id: Ibd2e0f436b7b37bff3f6de38c2f8013d4e7c8140
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/665858
Auto-Submit: John Stiles <johnstiles@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>