Move UploadInstance constructors from header
UploadInstance contains a vector BufferTextureCopyData which is forward
declared. The header also contains an explicitly defined default
constructor and implicitly defined move constructors. This prevents the
header from being built if BufferTextureCopyData is not defined in the
include chain. This CL fixes that by moving the constructor definitions
to the implementation source file.

Bug: b/276505403
Change-Id: Ief498de2d9503e6650334df83cfc48883c8cdc06
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/665639
Reviewed-by: Greg Daniel <egdaniel@google.com>
Commit-Queue: Sunny Sachanandani <sunnyps@google.com>