[graphite][compute] Support pre-compiled shader source
ComputeSteps can now set the `kSupportsNativeShader` flag to return a
shader module in the current Graphite backend's native shading language.
The ComputeStep must be able to provide a native shader source in all
backend formats that it will be run on. The returned shader source must
constitute a complete shader module with a valid compute kernel
entry-point function. The shader source is defined as an array of bytes
to accommodate binary formats (such as SPIR-V).

The main motivation for this feature is to support Vello shaders which
are pre-compiled from WGSL source and bundled in a library.

Bug: b/272520336
Change-Id: I4f32d2d8eae9cceba3bd4ce2ff08760e2209039e
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/666716
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Arman Uguray <armansito@google.com>