Add Fontations libs read-fonts, font-types via Bazel crates infra
Add a directory fontations/ to ports/ with a basic Rust test executable
which links against the read-fonts, font-types dependencies. The test
executable reads one of the test resource fonts and outputs the number
of glyphs in it, based on a call to the read_fonts parsing library.

Add a crates_repository() stanza to WORKSPACE.bazel for resolving the
dependencies that the fontations/ backend target needs.

Bug: skia:14243
Change-Id: Ib5bc4fae09a8d8aa0ad3a8cff1c4d8083deedba9
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/667257
Reviewed-by: Kevin Lubick <kjlubick@google.com>
Commit-Queue: Dominik Röttsches <drott@google.com>
Commit-Queue: Dominik Röttsches <drott@chromium.org>