Guard against negative weights in conic paths
The algorithm used for subdividing Bézier curves to
approximate a conic curve is from:
  "High order approximation of conic sections by
   quadratic splines" by Michael Floater

This algorithm assumes that the control point weight
will be greater than 0. The fuzzer is passing in a
negative value that causes the resulting subdivision
to produce control points that are too large.
If a negative weight is encountered, force it to 1.

Bug: skia:57716
Change-Id: If43335e686b641c0366420a9c79c2d7160806c3e
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/667436
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: Herb Derby <herb@google.com>