[graphite] Remove unnecessary text uploads.
Originally we were updating a Plot's upload token for every upload
request. However, we don't want to do this for the conditional uploads
we create at the start of a Recording -- if we do then each time we snap
a new Recording it will update the token and effectively consider those
uploads to take precedence over the ones that originally set up that
state. Then when we play the Recording back it will overwrite those
Plots even though they already contain the necessary glyphs. The Plots
should keep the token value for the non-conditional uploads that
originally set that state.

Bug: b/238791413
Change-Id: Id940535aa104e0a80f5e3fd65d28f711f2a6c793
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/666865
Commit-Queue: Jim Van Verth <jvanverth@google.com>
Reviewed-by: Herb Derby <herb@google.com>