Simplify conic chop
In CL 102425 code was added to handle overflow
conditions. An infinity result was checked, and the
calculation was redone using floats.

The new control points are calculated by summing a few
terms, and the result scaled by a factor >1. The
overflow to infinity happened during the summing of
the terms. Each of the terms is reduced by the same
scale factor which is <1. Therefore, applying the
scale factor before summing the terms will not
overflow to infinity.

The flag SK_SUPPORT_LEGACY_CONIC_CHOP was introduced
to control pixel differences in Chrome.

CL 102425 created the gm bigconic to force an overflow
in the conic subdivision code. This CL passes that
test.

Change-Id: I497d74a73951a9c8a0b835123b1bdc5a187a697b
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/667616
Reviewed-by: Jim Van Verth <jvanverth@google.com>
Commit-Queue: Herb Derby <herb@google.com>