Add dedicated SkRP op for broadcasting a constant.
Previously, we used the same op for uniforms and constants. This
forced us to track constants in a dedicated constant hash-map,
and also required an extra indirection at runtime to load the
constant value.

Now there is a dedicated op for writing a constant to a slot.

Change-Id: If3dac2070252f736cf0a423aefe081b33858eab6
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/667301
Commit-Queue: Jim Van Verth <jvanverth@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>
Reviewed-by: Jim Van Verth <jvanverth@google.com>