Deprecate SkImage::encodeToData and migrate all internal uses
By removing this method, we will further (completely?) decouple
SkImage from our encoders, allowing for them to not contribute
to code size when not needed.

This adds replacement functions to include/encode/Sk*Encoder.h.

This CL must land with the follow-on CL to preserve the ability
to use NDK encoders instead of bundling in ones based on libpng
et al.

Change-Id: Ie20a41cfdf65c975f640fb9368f363a588bd9394
Bug: skia:13983
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/667296
Reviewed-by: Herb Derby <herb@google.com>
Reviewed-by: Leon Scroggins <scroggo@google.com>