[graphite] Use replay translation in Dawn backend
Following the pattern used in Metal in http://review.skia.org/625730,
this applies the replay translation to the scissor and viewport. While
Metal only requires the scissor to be contained within the render target
dimensions, it seems Dawn requires this for the viewport as well, so the
application of the translation is a little different.

Bug: b/277611387
Change-Id: Ibea78bddece957a8f0def1f12e3abb7c63cb36d4
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/669196
Reviewed-by: Greg Daniel <egdaniel@google.com>
Commit-Queue: James Godfrey-Kittle <jamesgk@google.com>