Suppress failure of NDK encoding of kGray_8 images
It appears that before https://skia-review.googlesource.com/c/skia/+/667298,
we were shipping both the NDK-based encoders *and* the
normal encoders in our Android API 30+ tests (which set
skia_use_ndk_images). The encode_srgb tests explicitly called
the code path which used the latter.

After the above CL, the NDK-based encoders are the *only*
encoders shipped when skia_use_ndk_images is set. However,
it appears the NDK-based encoders can fail to encode
an image with the kGray_8 ColorType. Thus, we skip those,
as there is some disagreement between Skia and Android
for that color type.

Change-Id: I7f00e653812af8141767abeb6f547af7dfb9861d
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/669858
Reviewed-by: Leon Scroggins <scroggo@google.com>