Reject unsized arrays in function parameter lists.
These are already rejected when compiling a Runtime Effect, since
it's not legal in ES2, but were not detected as an error when
compiling to Metal/GLSL/etc.

A recent change (http://review.skia.org/670339) caused us to hit an
assertion when compiling such a shader; now we will detect it and
properly report an error in all modes.

Bug: oss-fuzz:57967
Change-Id: I7a0f7ffc105edfd6943c195c74eca48b7c589326
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/670676
Auto-Submit: John Stiles <johnstiles@google.com>
Commit-Queue: Kevin Lubick <kjlubick@google.com>
Reviewed-by: Kevin Lubick <kjlubick@google.com>