[graphite] Enforce std140 layout rule 4 for 2x2 matrices
std140 layout rule 4  states that all array elements must obey vec4
stride and alignment. For matrices this rule was correctly implemented
for row count 3 (via rule 3) but not for a row count of 2.

The std140 stride and alignment calculations now always round up to the
next multiple of 16 when handling matrix types, which is consistent with
SkSLMemoryLayout's implementation.

Bug: b/277588047

Change-Id: I02e8b84079ac512687494f4e6451d8c251d412d0
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/671416
Reviewed-by: John Stiles <johnstiles@google.com>
Commit-Queue: Arman Uguray <armansito@google.com>