Add Skrifa to the fontations third-party deps, test it in demo code
Skrifa was published on crates.io as well. Now we can use it as a crate
dependency here. Add that to the external/fontations description and
test it in the temporary test executable. Roll read-fonts and font-types
to 0.4.0 while we're at it.

Bug: skia:14260
Change-Id: I7f8648a5d4542b0224680f1e96654f2dff932413
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/671776
Auto-Submit: Dominik Röttsches <drott@chromium.org>
Reviewed-by: Kevin Lubick <kjlubick@google.com>
Commit-Queue: Kevin Lubick <kjlubick@google.com>