[graphite] Make ResourceCache be the owner of the ProxyCache
The reasoning here is that the ResourceCache will need access to the ProxyCache for its purge behavior and, to some extent, the ProxyCache
is just a specialization on top of the ResourceCache.

In this version the Context's ResourceCache doesn't get a ProxyCache - just the Recorders' ResourceCaches.

Bug: b/239604347
Change-Id: Ibb31095519bcd0422d73bbd805f5d3f1dc824c97
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/670240
Reviewed-by: Greg Daniel <egdaniel@google.com>
Commit-Queue: Robert Phillips <robertphillips@google.com>