Have SkFontMgr_FCI check the FCI passed to it
Add a release assert that the FontConfigInterface passed to the
constructor of SkFontMgr_FCI is not nullptr. This is a very inexpensive
check (normally happens once in a process' lifetime). If the
FontConfigInterface is nullptr then there will be issues much later when
the SkFontMgr_FCI is used. This change allows such problems to be caught
much earlier.

Change-Id: I4c7ff32f8196bd34ec9f119cde60f814803b9121
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/672156
Commit-Queue: Ben Wagner <bungeman@google.com>
Auto-Submit: Ben Wagner <bungeman@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: Brian Osman <brianosman@google.com>