Add SkJpegMetadataDecoder interface
Add a private SkJpegMetadataDecoder interface (similar to the adjacent
SkJpegGainmapEncoder interface). This interface can be used to extract
the gainmap from an encoded file.

Make the SkJpegMetadataDecoder::findGainmapImage function share the
implementation with the SkAndroidCodec path. To do this, change the
internal interface to no longer use jpeg_marker_struct directly
(since it can't be part of a public interface).

This interface will be updated to allow extracting the ICC profile and
EXIF data.

Bug: 1404000
Change-Id: I666c8a1e0df5847f4f289d8345edf0d3c30d5681
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/672256
Commit-Queue: Christopher Cameron <ccameron@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>