Disable an untested/unreachable GrSkSLFP optimization.
`GrSkSLFP::constantOutputForConstantInput` was previously able to
return a constant output color when given a constant input color,
even in the presence of child GrFragmentProcessors. However, the
child-evaluation code appears to be untested; I couldn't get a
breakpoint to trigger in it across all of dm. I also struggled to
make a test which reached it. (Very few GrFragmentProcessors
actually set the constant-output-for-constant-input flag!)

After working for a while on a hairy unit test, I realized that
this optimization isn't adding value (it never happens in real-
world usage) so it was wasted effort to bulletproof it. Instead,
we can just turn off the optimization flag in the presence of a
child GrFragmentProcessor. This also greatly simplifies the SkRP
implementation.

Change-Id: I4336ebad5852c308adce66583ebf00be94847af8
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/673657
Reviewed-by: Brian Osman <brianosman@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>