Implement GrSkSLFP::constantOutputForConstantInput in SkRP.
We don't use SkFilterColorProgram here; instead we assemble and
evaluate a Raster Pipeline for the effect on the fly. (In practice,
the effect being evaluated will almost always be the shader from
`GrFragmentProcessor::MakeColor`. It is a single-line color filter
which ignores its argument and returns a uniform.)

Change-Id: Ie156464775ffa10dec3b42cde48ead78be1fe447
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/673716
Commit-Queue: John Stiles <johnstiles@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>