Always run color space math in full-float within Ganesh
With HDR in particular, some of the intermediate values can have
magnitudes that break down badly with half. This CL fixes some HDR bugs
on Android. I'm hoping that the performance impact is minimal, and it
lets us remove a bunch of complication around a caps bit.

NOTE: There is additional cleanup to happen, but I'm splitting this CL
in two parts, so we can cherry-pick this bit to Android U.

Bug: b/270744068
Change-Id: Ie16f4aa66e01c6d881fb522efd5fb8467887f826
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/674517
Commit-Queue: Brian Osman <brianosman@google.com>
Reviewed-by: John Reck <jreck@google.com>
Reviewed-by: Nolan Scobie <nscobie@google.com>
Commit-Queue: Nolan Scobie <nscobie@google.com>