Fix anisotropic subset check in GrTextureEffect
Anisotropic sampling is meant to be disabled when there's a subset
since the in-shader vs. HW wrap-mode logic is only executed when there
is no anisotropy. Anisotropic sampling currently relies on the HW
tiling, so tiling at a subset boundary is unsupported.

In that case, the intent was to downgrade from anisotropic to mipmap
linear or simply bilinear filtering. The check for if there was a
subset flipped the inner and outer rectangles. In the original code,
checking if the subset was not contained in the proxy bounds was
always true. This would not drop anisotropy and the texture effect
would skip setting up in-shader clamping and wrapping.

Change-Id: I87c42abb787e8b30c5d0988a5240f1cf970b0168
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/675376
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: Brian Osman <brianosman@google.com>
Auto-Submit: Michael Ludwig <michaelludwig@google.com>