[graphite][compute] Add ComputeStep definitions for Vello stages
Added ComputeStep definitions for Vello compute pipeline stages. This
includes the full Vello pipeline and will eventually be extended for
coverage mask-only permutations as well. Each stage and the data flow
between them has been documented at a high level.

Currently the data flow layout for each ComputeStep is declared
statically. This is because the vello shaders crate currently doesn't
provide information to associate resource bindings between individual
stages. When it eventually does, these declarations could be generated
dynamically based on library metadata.

The comments in the header describes the order in which the pipelines
should be executed. This outline, combined with the provided resource
slot definitions can be used to dynamically assemble a DispatchGroup
that executes the algorithm.

Bug: b/272520336
Change-Id: I3ee6fe81219358cf6c17552fad667c6022655a40
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/681138
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Arman Uguray <armansito@google.com>