[graphite] Redefine PrioritizeGpuReads as AccessPattern
The PrioritizeGpuReads::kYes and PrioritizeGpuReads::kNo enum values are
now redefined as AccessPattern::kGpuOnly and
AccessPattern::kHostVisible. This redefinition more clearly captures a
resource creator's intended access pattern. The backend implementation
can use this determine mappability or buffer usage flags as it did
before.

PrioritizeGpuReads::kNo semantically did not impose any constraints on
what the actual buffer type should be and could have been thought of as
a "don't care" flag. Whereas, AccessPattern::kHostVisible implies an
intention to map the buffer. This is consistent with how
PrioritizeGpuReads::kNo was actually used in practice, in which it was
used by the Vulkan and Metal backend implementations to determine
whether a buffer is required to be mappable.

Change-Id: Ice40a44222dc4248f24024a5082fdef77f2cda57
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/681876
Commit-Queue: Arman Uguray <armansito@google.com>
Reviewed-by: Michael Ludwig <michaelludwig@google.com>