Add difficult test cases for sk_doubles_nearly_equal_ulps
Add test that probe the logical edges of
sk_doubles_nearly_equal_ulps. It highlights some problems
with the existing implementation.

In particular, it shows that big values and infinity are
equal, infinity and NaN are equal, and big values and NaN
are equal.

Change-Id: I276a41ea2f18d4be83a3f2ebb60e058396947bba
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/682257
Reviewed-by: Jim Van Verth <jvanverth@google.com>
Commit-Queue: Herb Derby <herb@google.com>